bufferered jpeg display

+++ src/framebuffer.cc
@@ -670,7 +670,7 @@ int write_fb_jpeg_to_surface(uint8_t *surface_data, const char *jpeg_file, uint8
             int g = img[i + 1]>>2;
             int r = img[i + 0]>>3;
             uint16_t t = r<<11 | g << 5 | b;
-            surface_pixels[ox+x_pos][oy+y_pos] = t;
+            surface_pixels[(int)(ox+x_pos)][(int)(oy+y_pos)] = t;
             //*((uint16_t*)(surface_data + location)) = t;
         }
     }
